Refactor: Button 컴포넌트 2차 리팩토링 완료
- 기존에 스니펫 지원 안되는 이슈 해결했습니다.
- children에 적용되는 font 관련 prop들을 삭제하였습니다.
- Button 컴포넌트 관련 폴더 구조를 변경하였습니다.
