+++ pages/test.tsx
@@ -1,5 +1,8 @@
 import styled from 'styled-components';
 import TextArea from '@/components/common/TextArea';
+import ButtonBase from '@/components/common/Button/ButtonBase';
+import { FONT_15, FONT_24 } from '@/styles/FontStyles';
+import Button from '@/components/common/Button';
 
 /**
  * 컴포넌트 실험용 페이지입니다. (컴포넌트 구현 완료 후 삭제 예정)
@@ -7,16 +10,18 @@ import TextArea from '@/components/common/TextArea';
 
 export default function Test() {
   return (
-    <Div>
-      <TextArea />
-    </Div>
+    <>
+      <Button.Add roundSize="L">냐냐</Button.Add>
+      <Button.Arrow type="left" isNotActive></Button.Arrow>
+      <Button.Arrow type="right"></Button.Arrow>
+      <Button.Plain style="primary" roundSize="L" isNotActive>
+        <Div>primary 버튼</Div>
+      </Button.Plain>
+    </>
   );
 }
 
 const Div = styled.div`
-  margin-bottom: 30px;
-  padding: 30px;
-  display: flex;
-  flex-direction: column;
-  gap: 80px;
+  color: white;
+  ${FONT_24};
 `;
